Some minor improvements + Punishment add action handler and packets.

+++ src/main/java/net/zylesh/dystellarcore/listeners/Scoreboards.java
@@ -4,6 +4,7 @@ import fr.mrmicky.fastboard.FastBoard;
 import net.luckperms.api.LuckPermsProvider;
 import net.minecraft.server.v1_7_R4.EntityPlayer;
 import net.zylesh.dystellarcore.DystellarCore;
+import net.zylesh.dystellarcore.core.User;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.craftbukkit.v1_7_R4.entity.CraftPlayer;
@@ -23,7 +24,7 @@ import java.util.concurrent.TimeUnit;
 
 public class Scoreboards implements Listener {
 
-    // TODO handle skywars disabling or enabling scoreboard, btw finish configuration guis too.
+    // TODO handle skywars disabling or enabling scoreboard.
 
     private static final ScheduledExecutorService threadPool = Executors.newSingleThreadScheduledExecutor();
     private static final Map<UUID, Map.Entry<FastBoard, Future<?>>> scoreboards = new HashMap<>();
@@ -34,31 +35,40 @@ public class Scoreboards implements Listener {
 
     @EventHandler(priority = EventPriority.HIGHEST)
     public void onJoin(PlayerJoinEvent event) {
-        FastBoard board = new FastBoard(event.getPlayer());
-        board.updateTitle(DystellarCore.SCOREBOARD_TITLE);
-        EntityPlayer player = ((CraftPlayer) event.getPlayer()).getHandle();
-        Future<?> task = threadPool.scheduleAtFixedRate(() -> {
-            List<String> updated = new ArrayList<>();
-            for (String s : DystellarCore.SCOREBOARD_LINES) {
-                updated.add(ChatColor.translateAlternateColorCodes('&', s.replaceFirst("<player_name>", event.getPlayer().getName()).replaceFirst("<ping>", String.valueOf(player.ping)).replaceFirst("<rank>", getRank(event.getPlayer())).replaceFirst("<online>", String.valueOf(Bukkit.getOnlinePlayers().size()))));
-            }
-            board.updateLines(updated);
-        }, DystellarCore.REFRESH_RATE_SCORE, DystellarCore.REFRESH_RATE_SCORE, TimeUnit.MILLISECONDS);
-        scoreboards.put(event.getPlayer().getUniqueId(), new AbstractMap.SimpleEntry<>(board, task));
+        User user =  User.get(event.getPlayer());
+        if (user.isScoreboardEnabled()) {
+            FastBoard board = new FastBoard(event.getPlayer());
+            board.updateTitle(DystellarCore.SCOREBOARD_TITLE);
+            EntityPlayer player = ((CraftPlayer) event.getPlayer()).getHandle();
+            Future<?> task = threadPool.scheduleAtFixedRate(() -> {
+                List<String> updated = new ArrayList<>();
+                for (String s : DystellarCore.SCOREBOARD_LINES) {
+                    updated.add(ChatColor.translateAlternateColorCodes('&', s.replaceFirst("<player_name>", event.getPlayer().getName()).replaceFirst("<ping>", String.valueOf(player.ping)).replaceFirst("<rank>", getRank(event.getPlayer())).replaceFirst("<online>", String.valueOf(Bukkit.getOnlinePlayers().size()))));
+                }
+                board.updateLines(updated);
+            }, DystellarCore.REFRESH_RATE_SCORE, DystellarCore.REFRESH_RATE_SCORE, TimeUnit.MILLISECONDS);
+            scoreboards.put(event.getPlayer().getUniqueId(), new AbstractMap.SimpleEntry<>(board, task));
+        }
     }
 
     @EventHandler
     public void onQuit(PlayerQuitEvent event) {
-        scoreboards.get(event.getPlayer().getUniqueId()).getKey().delete();
-        scoreboards.get(event.getPlayer().getUniqueId()).getValue().cancel(true);
-        scoreboards.remove(event.getPlayer().getUniqueId());
+        User user = User.get(event.getPlayer());
+        if (user.isScoreboardEnabled() && scoreboards.containsKey(event.getPlayer().getUniqueId())) {
+            scoreboards.get(event.getPlayer().getUniqueId()).getKey().delete();
+            scoreboards.get(event.getPlayer().getUniqueId()).getValue().cancel(true);
+            scoreboards.remove(event.getPlayer().getUniqueId());
+        }
     }
 
     @EventHandler
     public void onKick(PlayerKickEvent event) {
-        scoreboards.get(event.getPlayer().getUniqueId()).getKey().delete();
-        scoreboards.get(event.getPlayer().getUniqueId()).getValue().cancel(true);
-        scoreboards.remove(event.getPlayer().getUniqueId());
+        User user = User.get(event.getPlayer());
+        if (user.isScoreboardEnabled() && scoreboards.containsKey(event.getPlayer().getUniqueId())) {
+            scoreboards.get(event.getPlayer().getUniqueId()).getKey().delete();
+            scoreboards.get(event.getPlayer().getUniqueId()).getValue().cancel(true);
+            scoreboards.remove(event.getPlayer().getUniqueId());
+        }
     }
 
     private static String getRank(Player p) {
